Move simple security to libp2p/security
